feat: moved the officescardcomponent to a folder to make it more organized
